finished splitting out crud
