+++ api/app/tests/models/test_most_recent_historic_predictions.py
@@ -60,7 +60,7 @@ def mock_session(monkeypatch, data):
 
         monkeypatch.setattr(
             app.db.database, 'get_read_session', mock_get_session)
-        monkeypatch.setattr(app.db.crud, 'get_historic_station_model_predictions',
+        monkeypatch.setattr(app.db.crud.weather_models, 'get_historic_station_model_predictions',
                             mock_get_weather_station_model_predictions)
